requirements.txt
numpy==1.26.4
pandas==2.1.4
psycopg2-binary==2.9.9
pyarrow==17.0.0
pytest==8.3.3
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-toolbelt==1.0.0
SQLAlchemy==1.4.49
SQLAlchemy-JSONField==1.0.1.post0
sqlalchemy-redshift==0.8.14
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
urllib3==2.2.3